Extract Discover result stage
F3.4 — render-faithful decomposition of the high-risk result stage (StagePick)
out of Discover.jsx. No redesign, copy, styling, queue, alternates, writes,
ranking, audio, ceremony, parallax, fallback, or trailer-behaviour change — every
F3.1 finding stays deferred so the later one-pick redesign is reviewable.

Extracted (VERBATIM bodies; only export keywords + adapter props/hook-call added):
- resultPresentation.js — moodFilter (pure poster-filter mapping).
- hooks/useCountUp.js, hooks/useStreamingProvider.js (TMDB region/priority + null
  behaviour unchanged), hooks/useDiscoverResultActions.js (the Save/Mark-Watched/
  Skip/See-More handlers + savedState/watchedState + the mark-watched timer).
- sections/StreamingChip.jsx, sections/TrailerModal.jsx (YT singleton + portal +
  Escape + scroll-lock + overlay-close, no focus trap added), sections/AlternateCard.jsx,
  sections/StagePick.jsx.

Write/action behaviour preserved EXACTLY: handleSaveForLater → user_watchlist insert
+ updateImpression('saved') + 23505-as-success; handleMarkWatched → user_history
insert (source 'discover_marked') + updateImpression('watched') + the 600ms hold;
handleSkip (renamed from handleNotTonight, behaviour identical) → updateImpression
('skipped') + trackInteraction('dismiss', {action:'not_tonight', moods, intention,
energy, who}); handleSeeMore → updateImpression('clicked') + trackInteraction('click')
+ navigate('/movie/:tmdbId'); MAX_PICKS=15, ALTERNATES_VISIBLE=2, hidden-queue,
selected-top, sessionShownIds mutation, the reveal logSurfaceImpressions, and the
exhausted state all unchanged. No table name, field, event name, source value, or
metadata key changed.

Why a reducer / product redesign were deferred: this phase only relocates the result
surface so the one-pick redesign (remove alternates/queue/false-precision ring/
spectacle) can be done in isolated, reviewable steps. No reducer introduced. FFAudio,
AudioToggle, TMDB, DIARY_QUOTES, FILMS_FALLBACK, commitDiscoverPreferences, and
DiscoverBody remain in Discover.jsx; the AudioToggle reaches StagePick as an element
prop (FFAudio stays put). Discover.jsx changed only by import trims + the StagePick
import + the call-site audioToggle + the block removal (1349→508 lines).

Mechanical equivalence: every extracted body diffs against the pre-extraction
Discover.jsx as ONLY sanctioned changes — the actions-hook handler bodies and
StagePick's kept body are BYTE-IDENTICAL (verified by reverse-mapping vs `main`);
the 6 helpers differ only by the export prefix.

Tests: __tests__/DiscoverResult.test.jsx (26) — isolated, mocked Supabase/TMDB/auth/
nav/services. StagePick render (title/because/synopsis/taste-chip/mood-ring/alternates/
queue-count/exhausted/Start-Over/Tweak/alternate-promote); queue+actions (Skip+
MarkWatched promote, Save state, sessionShownIds, See-More nav+logs, no-nav without
tmdbId); the writes (watchlist + 23505 + error, history, skip impression+metadata,
no-write without user); StreamingChip priority labels; TrailerModal portal/focus/
Escape/overlay/inside-click; useCountUp init+cleanup. Stable 3×.

Untouched: derive.js, constants.js, the 6 pre-result sections, useDiscoverData.jsx,
discover.css, shared services, onboarding, /home, /movie/:id, router, auth, schema,
RLS, migrations, snapshots. No live Discover write triggered (mocked tests only).
