+++ src/features/discover/Discover.jsx
@@ -1,18 +1,13 @@
 import { useState, useMemo, useEffect, useRef, useCallback } from 'react'
-import { createPortal } from 'react-dom'
-import { useLocation, useNavigate } from 'react-router-dom'
-import { motion, AnimatePresence, useReducedMotion } from 'framer-motion'
+import { useLocation } from 'react-router-dom'
+import { motion, useReducedMotion } from 'framer-motion'
 import { supabase } from '@/shared/lib/supabase/client'
 import { useAuthSession } from '@/shared/hooks/useAuthSession'
 import { usePageMeta } from '@/shared/hooks/usePageMeta'
-import { scoreMovieForUser, logSurfaceImpressions, updateImpression } from '@/shared/services/recommendations'
+import { scoreMovieForUser } from '@/shared/services/recommendations'
 import { computeMatchPercent } from '@/shared/services/matchScore'
-import { trackInteraction } from '@/shared/services/interactions'
-import { getMovieWatchProviders } from '@/shared/api/tmdb'
-import { HP_GRAD } from '@/shared/lib/tokens'
-import { Play, SkipForward, X, Check, Bookmark } from 'lucide-react'
 import { DiscoverDataProvider, useDiscoverData } from './useDiscoverData'
-import { MOODS, ONBOARDING_TO_DISCOVER, constellationName, buildBecauseLine, diversifyTop3, predictDiscoverDefaults } from './derive'
+import { MOODS, ONBOARDING_TO_DISCOVER, diversifyTop3, predictDiscoverDefaults } from './derive'
 import { HP, TIME_OPTIONS } from './constants'
 import StageHero from './sections/StageHero'
 import StageMood from './sections/StageMood'
@@ -20,6 +15,7 @@ import StageNightStacked from './sections/StageNightStacked'
 import StageBreath from './sections/StageBreath'
 import StageReveal from './sections/StageReveal'
 import StageTitleCard from './sections/StageTitleCard'
+import StagePick from './sections/StagePick'
 import './discover.css'
 
 // /discover v5 — Magazine + the immersion senses
@@ -163,848 +159,6 @@ function Starfield({ tint }) {
 }
 
 // Internal Nav removed — AppShell provides the global TopNav.
-
-function useCountUp(target, duration=1400, delay=200) {
-  const [v, setV] = useState(0);
-  useEffect(() => {
-    let raf, start;
-    const t0 = setTimeout(() => {
-      const tick = (now) => { if (!start) start = now; const elapsed = now - start; const p = Math.min(elapsed / duration, 1); const eased = 1 - Math.pow(1 - p, 3); setV(Math.round(target * eased)); if (p < 1) raf = requestAnimationFrame(tick); };
-      raf = requestAnimationFrame(tick);
-    }, delay);
-    return () => { clearTimeout(t0); if (raf) cancelAnimationFrame(raf); };
-  }, [target, duration, delay]);
-  return v;
-}
-
-// buildWhyNow removed: the "Why now" box that explained the engine's
-// reasoning per intention/mood was one of 4 reason-boxes contributing to
-// the OCD/dyslexia clutter. The single "Why this one" proof line above the
-// action buttons now carries the reasoning, drawn from real engine signals.
-
-// buildPickProof removed: the multi-signal proof block (director hit, twin
-// rating, mood-axis fit, runtime fit, fallback) read as too much text
-// alongside the synopsis. Personalisation now lives in the visual signals
-// (match ring, mood chips, streaming chip, trailer button). If we bring
-// signal-language back later, the helper template is in git history.
-
-// MoodArc removed: the small SVG emotional-arc chart was one of 4 reason
-// surfaces contributing to the decision-page clutter. Lives on /movie/:id
-// (the Film File) for users who want the deep film-shape read.
-
-// ── Streaming provider chip (single best provider) ──────────────────────────
-// Borrowed from home-v2/sections-top.jsx — same compact chip with logo +
-// "Streaming on Netflix" / "Rent on Apple TV+" label. Returns null when TMDB
-// has no provider for the title; the chip then doesn't render.
-function useStreamingProvider(tmdbId) {
-  const [provider, setProvider] = useState(null);
-  useEffect(() => {
-    if (!tmdbId) { setProvider(null); return; }
-    const controller = new AbortController();
-    let cancelled = false;
-    setProvider(null);
-    getMovieWatchProviders(tmdbId, { region: 'CA', fallbackRegion: 'US', signal: controller.signal })
-      .then(data => {
-        if (cancelled) return;
-        const p = data?.providers?.[0];
-        if (p) setProvider(p);
-      })
-      .catch(() => { /* non-fatal */ });
-    return () => { cancelled = true; controller.abort(); };
-  }, [tmdbId]);
-  return provider;
-}
-
-function StreamingChip({ provider }) {
-  if (!provider) return null;
-  const label = provider.type === 'flatrate' ? 'Streaming on'
-    : provider.type === 'rent' ? 'Rent on'
-    : 'Buy on';
-  return (
-    <div style={{ display:'inline-flex', alignItems:'center', gap:10, padding:'8px 12px 8px 8px', borderRadius:8, background:'rgba(255,255,255,0.04)', border:`1px solid ${HP.border}`, maxWidth:'100%' }}>
-      <img
-        src={`https://image.tmdb.org/t/p/w92${provider.logoPath}`}
-        alt={provider.name}
-        style={{ width:28, height:28, borderRadius:5, objectFit:'cover', flex:'none' }}
-        loading="lazy"
-      />
-      <div style={{ minWidth:0 }}>
-        <div style={{ fontSize:9, fontWeight:600, letterSpacing:'0.18em', textTransform:'uppercase', color:HP.textFaint, fontFamily:'Outfit', lineHeight:1 }}>{label}</div>
-        <div style={{ fontSize:12, fontWeight:600, color:HP.text, fontFamily:'Outfit', lineHeight:1.2, marginTop:3 }}>{provider.name}</div>
-      </div>
-    </div>
-  );
-}
-
-// ── YouTube IFrame Player API — singleton loader ────────────────────────────
-// Loads https://www.youtube.com/iframe_api once (idempotent). The API
-// requires a global `onYouTubeIframeAPIReady` callback that fires after the
-// script finishes; we wrap it in a memoized Promise so multiple TrailerModal
-// mounts share the same load. Returns a Promise that resolves to window.YT.
-let _ytApiPromise = null;
-function loadYouTubeApi() {
-  if (_ytApiPromise) return _ytApiPromise;
-  _ytApiPromise = new Promise((resolve) => {
-    if (typeof window === 'undefined') return resolve(null);
-    if (window.YT && window.YT.Player) return resolve(window.YT);
-    const existing = window.onYouTubeIframeAPIReady;
-    window.onYouTubeIframeAPIReady = () => {
-      if (typeof existing === 'function') existing();
-      resolve(window.YT);
-    };
-    if (!document.querySelector('script[data-yt-iframe-api]')) {
-      const tag = document.createElement('script');
-      tag.src = 'https://www.youtube.com/iframe_api';
-      tag.async = true;
-      tag.dataset.ytIframeApi = '1';
-      document.head.appendChild(tag);
-    }
-  });
-  return _ytApiPromise;
-}
-
-// ── Inline trailer modal (YouTube IFrame Player) ────────────────────────────
-// Uses the official IFrame Player API (not a raw <iframe>) so we can listen
-// for the ENDED state and auto-close. That keeps the user on /discover after
-// the trailer instead of staring at YouTube's end-card with related-video
-// thumbnails (a brand leak). Click-on-overlay also closes — common modal
-// pattern users expect alongside Esc + the explicit X button.
-function TrailerModal({ open, youtubeKey, title, onClose }) {
-  const closeBtnRef = useRef(null);
-  const playerRef = useRef(null);
-  const playerSlotRef = useRef(null);
-  // Latest onClose held in a ref so the player effect's deps stay stable —
-  // otherwise the parent's inline `() => setTrailerOpen(false)` would
-  // re-create the player on every parent re-render.
-  const onCloseRef = useRef(onClose);
-  useEffect(() => { onCloseRef.current = onClose; }, [onClose]);
-
-  // Body scroll lock + Escape + initial focus
-  useEffect(() => {
-    if (!open) return;
-    const prevOverflow = document.body.style.overflow;
-    document.body.style.overflow = 'hidden';
-    const onKey = (e) => { if (e.key === 'Escape') onCloseRef.current?.(); };
-    window.addEventListener('keydown', onKey);
-    const t = setTimeout(() => closeBtnRef.current?.focus(), 0);
-    return () => {
-      window.removeEventListener('keydown', onKey);
-      document.body.style.overflow = prevOverflow;
-      clearTimeout(t);
-    };
-  }, [open]);
-
-  // Mount the YouTube player + listen for ENDED
-  useEffect(() => {
-    if (!open || !youtubeKey) return;
-    let cancelled = false;
-    loadYouTubeApi().then((YT) => {
-      if (cancelled || !YT || !playerSlotRef.current) return;
-      // new YT.Player replaces the slot div with an iframe sized to the
-      // parent (which is our 16/9 wrapper).
-      playerRef.current = new YT.Player(playerSlotRef.current, {
-        videoId: youtubeKey,
-        playerVars: {
-          autoplay: 1,
-          rel: 0,            // limit related videos to same channel
-          modestbranding: 1, // smaller YouTube logo
-          playsinline: 1,    // iOS inline play (no fullscreen takeover)
-        },
-        events: {
-          onStateChange: (e) => {
-            // YT.PlayerState.ENDED === 0. Hardcoded to avoid an extra
-            // import; the constant is stable in the YouTube API.
-            if (e.data === 0) onCloseRef.current?.();
-          },
-        },
-      });
-    });
-    return () => {
-      cancelled = true;
-      if (playerRef.current && typeof playerRef.current.destroy === 'function') {
-        try { playerRef.current.destroy(); } catch { /* ignore */ }
-      }
-      playerRef.current = null;
-    };
-  }, [open, youtubeKey]);
-
-  if (!open || !youtubeKey) return null;
-
-  // Overlay click → close (only when target is the overlay itself, not the
-  // inner iframe wrapper — stopPropagation on the wrapper prevents clicks
-  // inside the player from triggering close).
-  const onOverlayClick = (e) => { if (e.target === e.currentTarget) onClose(); };
-
-  // Portal to <body> so the modal escapes any parent stacking context.
-  // AppShell's main container creates a z:10 stacking context which would
-  // otherwise trap the modal beneath the z:50 fixed header — the close X
-  // was rendering but hidden behind the header bar. Portaling to body lifts
-  // the modal to the root stacking layer where z:300 actually means z:300.
-  return createPortal(
-    // Overlay click → close is a standard modal-dismiss pattern. Keyboard
-    // equivalent is the Escape key (handled in the effect above) + the
-    // explicit X close button (auto-focused on mount). The two a11y rules
-    // disabled here flag the click-on-overlay convention; suppress for
-    // this single modal — the dialog role + Escape handler cover the
-    // accessibility intent.
-    // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/click-events-have-key-events
-    <div
-      role="dialog" aria-modal="true" aria-label={`${title} trailer`}
-      onClick={onOverlayClick}
-      style={{ position:'fixed', inset:0, zIndex:300, background:'rgba(0,0,0,0.92)', backdropFilter:'blur(20px)', display:'flex', alignItems:'center', justifyContent:'center', padding:'24px', cursor:'pointer' }}
-    >
-      {/* Close X — bumped to opaque dark-glass so it reads against any
-         video frame (the older near-transparent treatment vanished
-         against bright content). Solid black background + brighter
-         white outline + slightly larger touch target. */}
-      <button
-        ref={closeBtnRef}
-        type="button"
-        onClick={onClose}
-        aria-label="Close trailer"
-        className="ff-trailer-close"
-        style={{ position:'absolute', top:20, right:20, width:44, height:44, borderRadius:999, background:'rgba(0,0,0,0.75)', border:'1px solid rgba(255,255,255,0.28)', color:'#fff', cursor:'pointer', display:'flex', alignItems:'center', justifyContent:'center', backdropFilter:'blur(6px)', transition:'background 0.2s ease, border-color 0.2s ease, transform 0.2s ease' }}
-      >
-        <X size={22} strokeWidth={2.2} />
-      </button>
-      {/* Esc-to-close hint — fades in for ~3s then out. Surfaces the
-         keyboard shortcut once on open (common in fullscreen video
-         players like YouTube + Netflix) without lingering. Pure
-         decoration once dismissed; sub-100ms CPU cost. */}
-      <div
-        aria-hidden="true"
-        style={{ position:'absolute', bottom:24, left:'50%', transform:'translateX(-50%)', zIndex:2, fontSize:11, fontFamily:'Outfit', fontWeight:500, color:'rgba(255,255,255,0.6)', letterSpacing:'0.06em', opacity:0, animation:'ff-titlecard 3.6s ease forwards', pointerEvents:'none', whiteSpace:'nowrap' }}
-      >
-        Press <span style={{ padding:'1px 6px', borderRadius:4, background:'rgba(255,255,255,0.1)', border:'1px solid rgba(255,255,255,0.22)', fontFamily:'ui-monospace, SFMono-Regular, Menlo, monospace', fontSize:10, marginLeft:4, marginRight:4, color:'#fff' }}>Esc</span> to close
-      </div>
-      {/* stopPropagation prevents overlay close when clicking inside the
-         iframe area. Pure click-intercept, no keyboard action expected. */}
-      {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */}
-      <div
-        onClick={(e) => e.stopPropagation()}
-        style={{ position:'relative', width:'100%', maxWidth:1080, aspectRatio:'16/9', borderRadius:8, overflow:'hidden', boxShadow:'0 24px 60px -16px rgba(0,0,0,0.8)', cursor:'default' }}
-      >
-        {/* Slot div — YT.Player replaces this with the actual iframe.
-           Sized to fill the 16/9 wrapper above. */}
-        <div ref={playerSlotRef} style={{ width:'100%', height:'100%' }} />
-      </div>
-    </div>,
-    document.body
-  );
-}
-
-function moodFilter(moodId) {
-  if (moodId === 'tense') return { cardFilter:'saturate(0.88)', overlay:'repeating-linear-gradient(0deg, rgba(239,68,68,0.04) 0px, rgba(239,68,68,0.04) 1px, transparent 1px, transparent 3px)' };
-  if (moodId === 'cozy')  return { overlay:'radial-gradient(ellipse at center, transparent 50%, rgba(0,0,0,0.4) 100%)' };
-  if (moodId === 'mythic')return { overlay:'radial-gradient(circle at 30% 20%, rgba(255,255,255,0.08), transparent 30%), radial-gradient(circle at 75% 75%, rgba(255,255,255,0.06), transparent 25%)' };
-  if (moodId === 'slow')  return { kenBurns:true };
-  if (moodId === 'bittersweet') return { cardFilter:'sepia(0.15) saturate(0.95)' };
-  if (moodId === 'cerebral')    return { cardFilter:'hue-rotate(-6deg) saturate(0.95)' };
-  if (moodId === 'tender')      return { halo:true };
-  return {};
-}
-
-// ── Alternate card — compact preview of an alternate pick. Click promotes
-// it to the top slot (the page crossfades and this card swaps places with
-// whatever was currently top). Title + meta + because-line; no match %,
-// no action buttons — those belong only on the focused top pick. Keeping
-// the alternates lightweight matches the "pick between three" intent
-// without multiplying decision load.
-function AlternateCard({ film, profile, selected, onPick }) {
-  const because = buildBecauseLine({ film, profile, selected, isAlt: true });
-  return (
-    <button
-      type="button"
-      onClick={onPick}
-      className="ff-alt-card"
-      style={{
-        display:'grid', gridTemplateColumns:'auto 1fr', gap:14,
-        padding:14, borderRadius:12,
-        background:HP.surface, border:`1px solid ${HP.border}`,
-        textAlign:'left', cursor:'pointer',
-        transition:'background 0.2s ease, border-color 0.2s ease, transform 0.2s ease',
-      }}
-    >
-      <img
-        src={film.poster}
-        alt={film.title}
-        loading="lazy"
-        style={{ width:84, height:126, objectFit:'cover', borderRadius:6, display:'block' }}
-      />
-      <div style={{ minWidth:0, display:'flex', flexDirection:'column', justifyContent:'space-between', padding:'2px 0' }}>
-        <div>
-          <h3 style={{ fontFamily:'Outfit', fontSize:16, fontWeight:500, color:HP.text, margin:0, letterSpacing:'-0.015em', lineHeight:1.2 }}>
-            {film.title}
-          </h3>
-          <div style={{ marginTop:6, fontFamily:'Outfit', fontSize:11, color:HP.textMuted, letterSpacing:'0.03em' }}>
-            {film.dir} &middot; {film.year} &middot; {film.runtime} min
-          </div>
-        </div>
-        {because && (
-          <p style={{ marginTop:8, marginBottom:0, fontFamily:'Outfit, Inter, sans-serif', fontSize:12, fontStyle:'italic', color:HP.textSoft, lineHeight:1.5, textWrap:'pretty' }}>
-            {because}
-          </p>
-        )}
-      </div>
-    </button>
-  );
-}
-
-// ── Stage 3 — Pick swiper (top pick + 2 alternates) ──
-function StagePick({ selected, who, energy, intention, results, profile, sessionShownIds, onRestart, onBack, blendHex }) {
-  // Session-only "hidden" set — when the user skips or marks watched, that
-  // film's id lands here and the visible picks filter it out. The next-best
-  // queued pick auto-advances into the top slot. Resets only when `results`
-  // changes (user tweaked inputs or started over). Pool depth is set by
-  // MAX_PICKS below — deep enough to feel like a real swipe deck without
-  // dropping into low-confidence territory.
-  const [hiddenTopIds, setHiddenTopIds] = useState(() => new Set());
-  // User can also manually pin one of the picks via the alternate cards
-  // ("pick from these" intent). null = use default (first non-hidden).
-  // Cleared on skip/watched so default takes over after dismissal.
-  const [selectedTopId, setSelectedTopId] = useState(null);
-  useEffect(() => { setHiddenTopIds(new Set()); setSelectedTopId(null); }, [results]);
-
-  // Pool of 15 picks — enough that the user can keep skipping without
-  // hitting exhausted prematurely, but capped before the engine starts
-  // returning weaker matches. 15 is a deep-enough swipe deck without
-  // becoming a feed.
-  const MAX_PICKS = 15;
-  // How many alternate cards to render below the focused top pick. The
-  // rest are "queued" — they auto-promote when the user skips/watches.
-  // 2 visible cards keeps the row scannable; deeper queue surfaces via
-  // the count in the section kicker.
-  const ALTERNATES_VISIBLE = 2;
-
-  const initialPicks = useMemo(() => results.slice(0, MAX_PICKS), [results]);
-  const visibleResults = useMemo(
-    () => initialPicks.filter(r => !hiddenTopIds.has(r.id)),
-    [initialPicks, hiddenTopIds]
-  );
-  const top = useMemo(() => {
-    if (selectedTopId) {
-      const pinned = visibleResults.find(r => r.id === selectedTopId);
-      if (pinned) return pinned;
-    }
-    return visibleResults[0];
-  }, [selectedTopId, visibleResults]);
-  // Visible alternates (rendered as cards) vs total remaining in the
-  // queue (rendered as a count in the kicker). When the user skips, the
-  // queue shrinks and the first queued film slides into the alternate
-  // slot. After all 15 are exhausted, falls through to the exhausted state.
-  const remainingAfterTop = useMemo(
-    () => visibleResults.filter(r => r.id !== top?.id),
-    [visibleResults, top]
-  );
-  const alternates = useMemo(
-    () => remainingAfterTop.slice(0, ALTERNATES_VISIBLE),
-    [remainingAfterTop]
-  );
-  const queuedCount = Math.max(0, remainingAfterTop.length - alternates.length);
-  const exhausted = !top;
-
-  const cName = constellationName(selected);
-  // Mood fit % — answers "how well does this film match what I asked for
-  // tonight?". Mean of the user's selected mood-axis fits against the
-  // film's fit vector, clamped 0-99. This is the headline ring number
-  // because the user JUST answered the mood question — the result page
-  // should reflect that answer. The taste % (top.match — calibrated
-  // baseScore via computeMatchPercent) is demoted to a meta-row chip,
-  // honoring both signals: "tonight" (mood-fit) and "you" (taste).
-  const moodFit = useMemo(() => {
-    if (!top?.fit || !selected || selected.length === 0) return null;
-    const sum = selected.reduce((a, id) => a + (top.fit[id] || 0), 0);
-    return Math.max(0, Math.min(99, Math.round((sum / selected.length) * 100)));
-  }, [top, selected]);
-  const matchAnim = useCountUp(moodFit ?? top?.match ?? 0, 1400, 250);
-  const [savedState, setSavedState] = useState('idle'); // idle | saving | saved | error
-  const [watchedState, setWatchedState] = useState('idle'); // idle | watched
-  const [trailerOpen, setTrailerOpen] = useState(false);
-  // Per-film Saved + Watched state — when the user clicks Save or Mark
-  // Watched, the button confirms success ("Saved", "Watched"). On
-  // auto-advance to a new film, both reset to idle so the new top can be
-  // saved/marked too.
-  useEffect(() => { setSavedState('idle'); setWatchedState('idle'); }, [top?.id]);
-  // Session-shown tracking — add every film that appears as top OR alt to
-  // the parent's sessionShownIds ref. diversifyTop3 reads this ref on
-  // each allResults recomputation (i.e., whenever user changes inputs)
-  // and demotes seen films so the next scenario surfaces fresh picks.
-  // Mutating a ref doesn't re-render — intentional; we only want the
-  // penalty applied on the NEXT input change, not within the current view.
-  useEffect(() => {
-    if (!sessionShownIds?.current) return;
-    if (top?.id) sessionShownIds.current.add(top.id);
-    for (const alt of (visibleResults || [])) {
-      if (alt?.id) sessionShownIds.current.add(alt.id);
-    }
-  }, [top?.id, visibleResults, sessionShownIds]);
-  const [mx, setMx] = useState(0); const [my, setMy] = useState(0);
-  const navigate = useNavigate();
-  const { user } = useAuthSession();
-  // Streaming provider (top 1 — flatrate first, then rent/buy). Hides cleanly
-  // when TMDB has nothing for this title.
-  const provider = useStreamingProvider(top?.tmdbId);
-  // "Because…" line — one signal that proves the engine knows the user.
-  // Recomputes whenever the top changes (manual pin or auto-advance).
-  const becauseLine = useMemo(
-    () => buildBecauseLine({ film: top, profile, selected }),
-    [top, profile, selected]
-  );
-  // Mood/tone chips: genre + the 2 strongest mood/tone tags. Tells the user
-  // what KIND of film this is at a glance, complementing the proof line
-  // (which says why it fits THEM). Capitalize + replace underscores so DB
-  // tags like "slow_burn" read as "Slow burn".
-  const chips = useMemo(() => {
-    if (!top) return [];
-    const out = [];
-    if (top.genre && top.genre !== 'Drama') out.push(top.genre); // skip plain "Drama" as too generic
-    else if (top.genre) out.push(top.genre);
-    const tags = [...(top._raw?.mood_tags || []), ...(top._raw?.tone_tags || [])]
-      .filter(Boolean)
-      .slice(0, 2)
-      .map(t => String(t).replace(/_/g, ' ').replace(/\b\w/g, c => c.toUpperCase()));
-    return [...out, ...tags].slice(0, 3);
-  }, [top]);
-  useEffect(() => {
-    const fn = (e) => { const cx = window.innerWidth/2, cy = window.innerHeight/2; setMx((e.clientX - cx)/cx); setMy((e.clientY - cy)/cy); };
-    window.addEventListener('mousemove', fn); return () => window.removeEventListener('mousemove', fn);
-  }, []);
-
-  // === Action handlers ===
-  // Stage 2 context shared across all interaction logs so the engine can
-  // learn from POSITIVE actions (save, watch, click), not just negative
-  // ones (skip). Previously Skip was the only handler writing
-  // user_interactions with mood/intention/energy/who — the engine had a
-  // negative bias because positive signals lost their full context.
-  const interactionContext = (action) => ({
-    movieId: top?.id,
-    source: 'discover',
-    metadata: { action, moods: selected, intention, energy, who },
-  });
-
-  // "See more" navigates to the deep film page (/movie/:tmdbId) where the
-  // user finds extended synopsis, providers, similar films, etc. Fires
-  // updateImpression('clicked') so the engine learns conversion AND
-  // trackInteraction('click', …) so the engine sees the full Stage 2
-  // context that drove the click — same metadata shape as Skip uses.
-  const handleSeeMore = () => {
-    if (!top?.tmdbId) return;
-    if (user?.id && top?.id) {
-      updateImpression(user.id, top.id, 'clicked').catch(() => {})
-      trackInteraction('click', interactionContext('see_more')).catch(() => {})
-    }
-    navigate(`/movie/${top.tmdbId}`);
-  };
-  const handleSaveForLater = async () => {
-    if (!user?.id || !top?.id || savedState !== 'idle') return;
-    const filmId = top.id;
-    setSavedState('saving');
-    try {
-      const { error } = await supabase
-        .from('user_watchlist')
-        .insert({ user_id: user.id, movie_id: filmId });
-      if (error && error.code !== '23505') throw error; // 23505 = unique violation (already in list)
-      setSavedState('saved');
-      // Engine learning: flip the impression's added_to_watchlist flag
-      // AND log the interaction with Stage 2 context. Both fire after
-      // the watchlist write succeeds so we don't credit a failed save.
-      updateImpression(user.id, filmId, 'saved').catch(() => {})
-      trackInteraction('save', interactionContext('save_for_later')).catch(() => {})
-    } catch (e) {
-      console.error('[Discover.saveForLater]', e);
-      setSavedState('error');
-    }
-  };
-  // Skip Tonight — auto-advances to the next pick instead of bouncing home.
-  // The page now behaves as a 3-pick swiper: skip the top → alternate #1
-  // becomes top → alternate #2 becomes top → exhausted state. Each skip
-  // still fires:
-  //   1. user_interactions ('dismiss', source 'discover') — analytics log
-  //   2. recommendation_impressions.skipped = true → engine's negative-
-  //      signal model penalises matching directors/genres next /discover run
-  // The skipped film also lands in hiddenTopIds for the rest of this
-  // session AND gets a 30-day hard exclusion via useDiscoverData's
-  // recent-skip query.
-  const handleNotTonight = () => {
-    if (!top?.id) return;
-    trackInteraction('dismiss', {
-      movieId: top.id,
-      source: 'discover',
-      metadata: {
-        action: 'not_tonight',
-        moods: selected,
-        intention,
-        energy,
-        who,
-      },
-    }).catch(() => {})
-    if (user?.id) {
-      updateImpression(user.id, top.id, 'skipped').catch(() => {})
-    }
-    setHiddenTopIds(prev => new Set([...prev, top.id]));
-    setSelectedTopId(null);
-  };
-
-  // Click an alternate card to promote it to top. AnimatePresence
-  // crossfades the spread; the alternate that was promoted swaps places
-  // with whatever was currently top. No-op if clicked the current top.
-  const handlePickAlternate = (altId) => {
-    if (!altId || altId === top?.id) return;
-    setSelectedTopId(altId);
-  };
-
-  // Mark Watched — user has already seen this film. Closes the loop both
-  // for the recommendation engine (don't surface this again) and the
-  // user's own library. The button flips to "Watched" for a beat of
-  // confirmation, then auto-advances to the next pick. Capture top.id in
-  // a closure so the setTimeout still fires against the right film even
-  // after the state change starts swapping in the next pick.
-  const markWatchedTimeoutRef = useRef(null);
-  useEffect(() => () => {
-    if (markWatchedTimeoutRef.current) clearTimeout(markWatchedTimeoutRef.current);
-  }, []);
-  const handleMarkWatched = async () => {
-    if (!top?.id || !user?.id || watchedState !== 'idle') return;
-    const filmId = top.id;
-    setWatchedState('watched');
-    try {
-      // 23505 (unique violation) = already in history; treat as success
-      const { error } = await supabase
-        .from('user_history')
-        .insert({ user_id: user.id, movie_id: filmId, source: 'discover_marked' });
-      if (error && error.code !== '23505') throw error;
-    } catch (e) {
-      console.error('[Discover.markWatched]', e);
-      // Even on insert failure, hide locally so the user can move on
-    }
-    updateImpression(user.id, filmId, 'watched').catch(() => {});
-    trackInteraction('watch', interactionContext('mark_watched')).catch(() => {});
-    // 600ms holds "Watched ✓" long enough to register as confirmation
-    // before the crossfade swap (180ms) carries the next pick in. Cleared
-    // on unmount so a quick Tweak inputs / Start over click during the
-    // hold doesn't fire a setState on an unmounted component.
-    markWatchedTimeoutRef.current = setTimeout(() => {
-      setHiddenTopIds(prev => new Set([...prev, filmId]));
-      setSelectedTopId(null);
-      markWatchedTimeoutRef.current = null;
-    }, 600);
-  };
-
-  // Stage 3 impression log — fires per current top pick. The page now
-  // behaves as a 3-pick swiper (skip → alt #1 becomes top → alt #2 becomes
-  // top → exhausted), so each new top deserves its own impression row
-  // under placement='discover'. Alternates are not logged as separate
-  // impressions until they're promoted to top — they're queued, not yet
-  // seen as a "pick" by the user. Per-day-deduped by the table's unique
-  // key, so re-visits don't inflate counts.
-  const topId = top?.id
-  useEffect(() => {
-    if (!user?.id || !topId || !top) return
-    logSurfaceImpressions({
-      userId: user.id,
-      films: [{ id: top.id, engineScore: top.match, _pickReasonLabel: 'discover_top_pick' }],
-      placement: 'discover',
-      pickReasonType: 'discover_reveal',
-      pickReasonLabel: `Discover · ${cName}`,
-    })
-    // top.id pins the impression — when the swiper auto-advances we log
-    // the new top as its own row.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [user?.id, topId])
-
-  const filter = moodFilter(selected[0]);
-
-  // Exhausted state — user skipped/watched all 3 initial picks. Editorial
-  // dead-end with two doors: tweak inputs (revisit Stage 2 — same
-  // constellation, different night-shape answers) or start over (back to
-  // Stage 0 — fresh moods). Matches the design language with eyebrow +
-  // italic accent + brand-gradient primary action. No new picks magic-
-  // refilled here; that would feel like infinite scroll, which the OCD/
-  // dyslexia cleanup is explicitly against.
-  if (exhausted) {
-    return (
-      <section className="ff-discover-section" style={{ position:'relative', minHeight:'70vh', display:'flex', flexDirection:'column', alignItems:'center', justifyContent:'center', padding:'80px 24px', textAlign:'center', animation:'ff-fade 0.5s ease' }}>
-        <AudioToggle />
-        <div style={{ fontSize:10, fontWeight:700, letterSpacing:'0.32em', textTransform:'uppercase', color:blendHex, marginBottom:18, fontFamily:'Outfit', display:'inline-flex', alignItems:'center', gap:10 }}>
-          <span style={{ height:1, width:22, background:blendHex, opacity:0.6 }} />
-          End of edition
-        </div>
-        <h2 style={{ fontFamily:'Outfit', fontSize:'clamp(36px, 5vw, 62px)', lineHeight:1.02, fontWeight:200, letterSpacing:'-0.04em', color:HP.text, margin:0, maxWidth:760, textWrap:'balance' }}>
-          That&rsquo;s <em style={{ fontStyle:'italic', fontWeight:300, color:HP.textSoft }}>everything</em> for tonight.
-        </h2>
-        <p style={{ marginTop:18, fontFamily:'Outfit, Inter, sans-serif', fontSize:14, color:HP.textMuted, fontStyle:'italic', maxWidth:480, lineHeight:1.6 }}>
-          Tweak what you told me &mdash; or start fresh. New picks are seconds away.
-        </p>
-        <div style={{ marginTop:32, display:'flex', gap:10, flexWrap:'wrap', justifyContent:'center' }}>
-          <button onClick={onBack} style={{ display:'inline-flex', alignItems:'center', justifyContent:'center', padding:'12px 22px', borderRadius:10, background:HP_GRAD, border:'none', color:'#fff', fontFamily:'Outfit', fontSize:13, fontWeight:700, letterSpacing:'0.01em', cursor:'pointer', boxShadow:'0 10px 26px -10px rgba(236,72,153,0.55)' }}>&larr; Tweak inputs</button>
-          <button onClick={onRestart} style={{ display:'inline-flex', alignItems:'center', justifyContent:'center', padding:'12px 18px', borderRadius:10, background:HP.surface, border:`1px solid ${HP.borderStrong}`, color:HP.text, fontFamily:'Outfit', fontSize:13, fontWeight:500, cursor:'pointer' }}>Start over</button>
-        </div>
-      </section>
-    );
-  }
-
-  // Synopsis — TMDB's real overview. Gracefully degrades: if the film row
-  // has no overview (rare; some films missing it, or fallback set), we hide
-  // the article paragraph entirely rather than fall back to the older
-  // templated ffTake prose, which was the same wording for every film with
-  // only the director swapped (read as curator voice but was a Mad Lib).
-  const synopsis = top.overview && top.overview.trim().length > 20 ? top.overview.trim() : null;
-
-  // Title words for typesetting
-  const titleWords = top.title.split(' ');
-
-  return (
-    <section className="ff-discover-section" style={{ position:'relative', animation:'ff-fade 0.7s ease' }}>
-      <AudioToggle />
-      {/* Film grain + vignette overlays */}
-      <div aria-hidden style={{ position:'fixed', inset:0, pointerEvents:'none', zIndex:2, opacity:0.045, mixBlendMode:'overlay', backgroundImage:'url("data:image/svg+xml;utf8,<svg xmlns=%22http://www.w3.org/2000/svg%22 width=%22220%22 height=%22220%22><filter id=%22n%22><feTurbulence type=%22fractalNoise%22 baseFrequency=%220.9%22 numOctaves=%222%22 stitchTiles=%22stitch%22/></filter><rect width=%22100%25%22 height=%22100%25%22 filter=%22url(%23n)%22/></svg>")', animation:'ff-grain 1.6s steps(6) infinite' }} />
-      <div aria-hidden style={{ position:'fixed', inset:0, pointerEvents:'none', zIndex:2, background:'radial-gradient(ellipse at center, transparent 50%, rgba(0,0,0,0.55) 100%)' }} />
-      {/* No mast banner — the user is already on /discover; the brand
-         chrome ("FeelFlick · The Discover Edition") + curator note
-         ("Tonight's film, for your constellation X") were two rows of
-         magazine theater that added cognitive load before the decision.
-         The constellation name still pays off in /movie/:tmdbId if the
-         user wants to read deep. */}
-
-      {/* Spread — two columns on desktop, single linear flow on mobile.
-         Wrapped in AnimatePresence keyed on top.id so that when the user
-         skips or marks Watched it, the current pick fades out and the
-         next pick fades in. Without this the swap was a jarring snap —
-         posters changed mid-page with no continuity. mode="wait" keeps
-         the layout stable (next pick doesn't mount until current finishes
-         exiting), and 180ms duration keeps total swap snappy (~360ms). */}
-      <AnimatePresence mode="wait" initial={false}>
-        <motion.div
-          key={top.id}
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={{ duration: 0.18, ease: 'easeOut' }}
-        >
-      <div className="ff-stage3-spread">
-        {/* LEFT PAGE — poster + match ring only. The film strip ribbon,
-           pairing notes, and IN THIS ISSUE TOC sidebar all gone — pure
-           magazine chrome with no decision payload. The spine gutter is
-           also gone (it framed a 2-page magazine; without the surrounding
-           chrome it now just marks an arbitrary divide). */}
-        <div className="ff-stage3-spread__left" style={{ position:'relative', transform:`translate(${mx*-6}px, ${my*-4}px)`, transition:'transform 0.4s cubic-bezier(.2,.7,.2,1)' }}>
-          <div style={{ position:'relative' }}>
-            <div style={{ position:'absolute', inset:-14, borderRadius:14, background:`radial-gradient(ellipse at center, ${blendHex}55, transparent 70%)`, filter:'blur(30px)', animation: filter.halo ? 'ff-bloom-pulse-strong 4s ease-in-out infinite' : 'ff-bloom-pulse 4s ease-in-out infinite' }} />
-            <div style={{ position:'relative', overflow:'hidden', borderRadius:8, boxShadow:`0 30px 60px -20px rgba(0,0,0,0.8), 0 0 0 1px ${blendHex}33` }}>
-              <img src={top.poster} alt={top.title} style={{ width:'100%', aspectRatio:'2/3', objectFit:'cover', display:'block', filter: filter.cardFilter || 'none', animation: filter.kenBurns ? 'ff-kenburns 14s ease-in-out infinite alternate' : 'ff-poster-in 1s cubic-bezier(.2,.7,.2,1)' }} />
-              {filter.overlay && <div style={{ position:'absolute', inset:0, background:filter.overlay, pointerEvents:'none', mixBlendMode:'overlay' }} />}
-            </div>
-            {/* Match ring */}
-            <div style={{ position:'absolute', right:-22, bottom:-22, width:104, height:104, borderRadius:999, background:`conic-gradient(${blendHex} ${matchAnim*3.6}deg, rgba(255,255,255,0.07) 0)`, display:'flex', alignItems:'center', justifyContent:'center', padding:5, boxShadow:`0 0 32px ${blendHex}66` }}>
-              <div style={{ width:'100%', height:'100%', borderRadius:999, background:HP.bgDeep, display:'flex', flexDirection:'column', alignItems:'center', justifyContent:'center' }}>
-                <div style={{ fontFamily:'Outfit', fontSize:26, fontWeight:300, color:HP.text, letterSpacing:'-0.04em', lineHeight:1, fontVariantNumeric:'tabular-nums' }}>{matchAnim}<span style={{ fontSize:11, color:HP.textMuted }}>%</span></div>
-                <div style={{ fontSize:7, color:HP.textMuted, letterSpacing:'0.2em', textTransform:'uppercase', fontFamily:'Outfit', marginTop:2 }}>Mood fit</div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* RIGHT PAGE */}
-        <div className="ff-stage3-spread__right" style={{ transform:`translate(${mx*-3}px, ${my*-2}px)`, transition:'transform 0.4s cubic-bezier(.2,.7,.2,1)' }}>
-          {/* Because-line — the one signal that proves the engine knows
-             the user. Tinted to the mood-blend hex so it reads as part
-             of the user's constellation, not engine chrome. Hides when
-             no honest signal exists (cold-start with no mood) rather
-             than printing filler. */}
-          {becauseLine && (
-            <div style={{ marginBottom:14, fontFamily:'Outfit, Inter, sans-serif', fontSize:13, fontStyle:'italic', fontWeight:400, color:blendHex, letterSpacing:'0.01em', textWrap:'balance', opacity:0, animation:'ff-fade 0.5s ease 0.15s forwards' }}>
-              {becauseLine}
-            </div>
-          )}
-          {/* Title (no kicker, no byline — the page IS the pick, no need to
-             pre-announce it). Word-by-word reveal animation kept; it's pure
-             motion, not content load. */}
-          <h1 style={{ fontFamily:'Outfit', fontSize:'clamp(44px, 5.6vw, 76px)', lineHeight:0.96, fontWeight:300, letterSpacing:'-0.045em', color:HP.text, margin:0, textWrap:'balance' }}>
-            {titleWords.map((w, i) => <span key={i} style={{ display:'inline-block', opacity:0, animation:`ff-word-in 0.55s cubic-bezier(.2,.7,.2,1) ${0.2 + i*0.08}s forwards`, marginRight:'0.2em' }}>{w}</span>)}
-          </h1>
-
-          {/* Meta — director · year · runtime · taste %. The taste % was
-             demoted from the headline ring (which now shows mood-fit, the
-             number that actually answers what the user just asked) into
-             this quiet meta row. Both signals stay visible: "tonight"
-             (mood-fit, big) and "you" (taste, small). textFaint colour
-             keeps it from competing with the director name. */}
-          <div style={{ marginTop:14, display:'flex', alignItems:'center', gap:10, flexWrap:'wrap', fontFamily:'Outfit', fontSize:13, color:HP.textMuted, letterSpacing:'0.04em', opacity:0, animation:'ff-fade 0.6s ease 0.7s forwards' }}>
-            <span>{top.dir}</span>
-            <span style={{ color:HP.textFaint }}>·</span>
-            <span>{top.year}</span>
-            <span style={{ color:HP.textFaint }}>·</span>
-            <span>{top.runtime} min</span>
-            {Number.isFinite(top.match) && (
-              <>
-                <span style={{ color:HP.textFaint }}>·</span>
-                <span style={{ color:HP.textFaint }} title="How well this film fits your stable taste profile">{top.match}% taste</span>
-              </>
-            )}
-          </div>
-
-          {/* Mood/tone chips — what KIND of film this is at a glance.
-             3 chips max: primary genre + 2 strongest mood/tone tags.
-             Complements the italic proof line below (which says why it
-             fits the USER) by saying what it IS. No background — borders
-             only, so visually quieter than the action buttons. */}
-          {chips.length > 0 && (
-            <div style={{ marginTop:14, display:'flex', flexWrap:'wrap', gap:7, opacity:0, animation:'ff-fade 0.6s ease 0.8s forwards' }}>
-              {chips.map((c, i) => (
-                <span key={`${c}-${i}`} style={{ display:'inline-flex', alignItems:'center', padding:'4px 10px', borderRadius:999, border:`1px solid ${HP.border}`, fontFamily:'Outfit', fontSize:11, fontWeight:500, color:HP.textSoft, letterSpacing:'0.02em' }}>
-                  {c}
-                </span>
-              ))}
-            </div>
-          )}
-
-          {/* No FOR YOU prose block — the proof lines (mood-axis hit,
-             runtime fit, fallback) read as "a lot to read" alongside the
-             synopsis. The match ring (calibrated %) + mood chips + trailer
-             button already convey personalisation visually. Deeper signal
-             readout lives on /movie/:tmdbId for users who want it. */}
-
-          {/* Synopsis — real TMDB overview, plain paragraph, no drop-cap.
-             Demoted below the FOR YOU block because it's generic info every
-             recommender has; the personal proof above is what makes this
-             feel like FeelFlick. Drop-cap was cut earlier (fragments
-             reading for dyslexic users). Max-width 580 caps line length
-             for readability. Hides when no overview exists. */}
-          {synopsis && (
-            <p style={{ marginTop:22, marginBottom:0, fontFamily:'Inter, sans-serif', fontSize:14.5, color:HP.textMuted, lineHeight:1.7, textWrap:'pretty', maxWidth:580, opacity:0, animation:'ff-fade 0.7s ease 1.2s forwards' }}>
-              {synopsis}
-            </p>
-          )}
-
-          {/* Streaming provider — one chip with logo + "Streaming on Netflix"
-             style label. Renders only when TMDB has a provider for this
-             title. Tells the user in one glance whether the recommendation
-             is even reachable tonight, removing a click of friction at the
-             most decision-critical moment. Detailed provider list (Stream
-             / Rent / Buy categories) lives on /movie/:tmdbId. */}
-          {provider && (
-            <div style={{ marginTop:18, opacity:0, animation:'ff-fade 0.7s ease 1.3s forwards' }}>
-              <StreamingChip provider={provider} />
-            </div>
-          )}
-
-          {/* Action buttons. Primary = gradient "See More" → /movie/:tmdbId
-             (deep film page; "Watch now" used to lie because the button
-             didn't play the film). Secondary row: Trailer (when YouTube
-             key exists), Save, Mark Watched, Skip for tonight. Save +
-             Mark Watched both flip to a confirmed state ("Saved" /
-             "Watched"); Mark Watched + Skip both auto-advance the swiper
-             (positive history write vs. negative dismiss signal). Style:
-             rounded-rectangle pills (12px radius) with icons on the left,
-             dark surface fill, bolder typography — same shape language as
-             the briefing action bar. */}
-          <div className="ff-stage3-actions" style={{ opacity:0, animation:'ff-fade 0.7s ease 1.4s forwards' }}>
-            <button
-              type="button"
-              onClick={handleSeeMore}
-              style={{ display:'inline-flex', alignItems:'center', justifyContent:'center', padding:'12px 22px', borderRadius:10, background:HP_GRAD, border:'none', color:'#fff', fontFamily:'Outfit', fontSize:13, fontWeight:700, letterSpacing:'0.01em', cursor:'pointer', boxShadow:'0 10px 26px -10px rgba(236,72,153,0.55)' }}
-            >
-              See More &rarr;
-            </button>
-            {top.trailerKey && (
-              <button
-                type="button"
-                onClick={() => setTrailerOpen(true)}
-                title="Watch the trailer"
-                style={{ display:'inline-flex', alignItems:'center', gap:8, padding:'12px 18px', borderRadius:10, background:HP.surface, border:`1px solid ${HP.borderStrong}`, color:HP.text, fontFamily:'Outfit', fontSize:13, fontWeight:500, cursor:'pointer' }}
-              >
-                <Play size={14} fill="currentColor" />
-                <span>Trailer</span>
-              </button>
-            )}
-            <button
-              type="button"
-              onClick={handleMarkWatched}
-              disabled={watchedState === 'watched'}
-              title="I've already seen this one"
-              style={{ display:'inline-flex', alignItems:'center', gap:8, padding:'12px 18px', borderRadius:10, background: watchedState === 'watched' ? `${HP.green}14` : HP.surface, border:`1px solid ${watchedState === 'watched' ? HP.green + '66' : HP.borderStrong}`, color: watchedState === 'watched' ? HP.green : HP.text, fontFamily:'Outfit', fontSize:13, fontWeight:500, cursor: watchedState === 'idle' ? 'pointer' : 'default' }}
-            >
-              <Check size={14} strokeWidth={2.5} />
-              <span>{watchedState === 'watched' ? 'Watched' : 'Mark Watched'}</span>
-            </button>
-            <button
-              type="button"
-              onClick={handleSaveForLater}
-              disabled={savedState === 'saving' || savedState === 'saved'}
-              style={{ display:'inline-flex', alignItems:'center', gap:8, padding:'12px 18px', borderRadius:10, background: savedState === 'saved' ? `${HP.green}14` : HP.surface, border:`1px solid ${savedState === 'saved' ? HP.green + '66' : HP.borderStrong}`, color: savedState === 'saved' ? HP.green : HP.text, fontFamily:'Outfit', fontSize:13, fontWeight:500, cursor: savedState === 'idle' ? 'pointer' : 'default' }}
-            >
-              <Bookmark size={14} fill={savedState === 'saved' ? 'currentColor' : 'none'} strokeWidth={2} />
-              <span>
-                {savedState === 'idle' && 'Save'}
-                {savedState === 'saving' && 'Saving…'}
-                {savedState === 'saved' && 'Saved'}
-                {savedState === 'error' && 'Try again'}
-              </span>
-            </button>
-            <button
-              type="button"
-              onClick={handleNotTonight}
-              title="Skip — not tonight"
-              style={{ display:'inline-flex', alignItems:'center', gap:8, padding:'12px 18px', borderRadius:10, background:HP.surface, border:`1px solid ${HP.border}`, color:HP.textSoft, fontFamily:'Outfit', fontSize:13, fontWeight:500, cursor:'pointer' }}
-            >
-              <SkipForward size={14} />
-              <span>Skip for tonight</span>
-            </button>
-          </div>
-          {/* No depth zone — the match ring + mood chips + streaming chip
-             carry the reasoning visually. The deeper magazine treatment
-             (diary callback, why-now, emotional arc, taste twin) was 4
-             boxes of competing chrome — OCD-hostile and dyslexia-hostile.
-             Users who want a deep read on the film have /movie/:tmdbId
-             waiting (linked from the See More button). */}
-        </div>
-      </div>
-
-      {/* Alternates row — two smaller cards for the other 2 picks from
-         this batch. Click to promote to top (crossfade swap). Each card
-         carries its own because-line so the user can see WHY each is
-         on the table. No match %, no action buttons — those belong only
-         to the focused top pick. Hides when no alternates are visible
-         (after skip/watched bring the count down to 1). */}
-      {alternates.length > 0 && (
-        <div style={{ marginTop:48, maxWidth:1080, marginLeft:'auto', marginRight:'auto', padding:'0 24px' }}>
-          <div style={{ fontSize:10, fontWeight:700, letterSpacing:'0.28em', textTransform:'uppercase', color:HP.textMuted, marginBottom:16, fontFamily:'Outfit', display:'inline-flex', alignItems:'center', gap:10 }}>
-            <span style={{ height:1, width:22, background:HP.textMuted, opacity:0.5 }} />
-            <span>Or pick from these</span>
-            {queuedCount > 0 && (
-              <span style={{ color:HP.textFaint, letterSpacing:'0.18em' }} title={`${queuedCount} more films queued — skip or mark watched to see them`}>
-                &middot; {queuedCount} more queued
-              </span>
-            )}
-          </div>
-          <div className="ff-stage3-alternates">
-            {alternates.map(alt => (
-              <AlternateCard
-                key={alt.id}
-                film={alt}
-                profile={profile}
-                selected={selected}
-                onPick={() => handlePickAlternate(alt.id)}
-              />
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Footer — Tweak inputs / Start over, the two ways back into the
-         flow without committing to the current pick. Quiet pills, not
-         primary buttons — the action gravity belongs on the spread above. */}
-      <div style={{ marginTop:56, paddingTop:24, borderTop:`1px solid ${HP.border}`, display:'flex', alignItems:'center', justifyContent:'center', gap:10, maxWidth:1080, margin:'56px auto 0' }}>
-        <button onClick={onBack}    style={{ padding:'9px 14px', borderRadius:8, background:'transparent', border:`1px solid ${HP.border}`, color:HP.textMuted, fontFamily:'Outfit', fontSize:11, fontWeight:500, cursor:'pointer' }}>← Tweak inputs</button>
-        <button onClick={onRestart} style={{ padding:'9px 14px', borderRadius:8, background:'transparent', border:`1px solid ${HP.border}`, color:HP.textMuted, fontFamily:'Outfit', fontSize:11, fontWeight:500, cursor:'pointer' }}>Start over</button>
-      </div>
-        </motion.div>
-      </AnimatePresence>
-
-      <TrailerModal open={trailerOpen} youtubeKey={top.trailerKey} title={top.title} onClose={() => setTrailerOpen(false)} />
-    </section>
-  );
-}
 
 // Commit the user's Stage 2 selections to user_discover_preferences —
 // increments the count for each selected option in each dimension, plus
@@ -1291,7 +445,7 @@ function DiscoverBody() {
         {stage === 2.3 && <StageBreath onDone={()=>setStage(2.5)} />}
         {stage === 2.5 && <StageReveal selected={selected.length>0?selected:['slow','tender']} onDone={()=>setStage(2.7)} />}
         {stage === 2.7 && <StageTitleCard title={(allResults[0]||{}).title || ''} onDone={()=>setStage(3)} playTitleCue={()=>FFAudio.chord()} />}
-        {stage === 3   && <StagePick selected={selected.length>0?selected:['slow','tender']} who={who} energy={energy} intention={intention} results={allResults} profile={profile} sessionShownIds={sessionShownIds} onRestart={()=>{ setStage(0); setSelected([]); }} onBack={()=>setStage(2)} blendHex={blendHex} />}
+        {stage === 3   && <StagePick selected={selected.length>0?selected:['slow','tender']} who={who} energy={energy} intention={intention} results={allResults} profile={profile} sessionShownIds={sessionShownIds} onRestart={()=>{ setStage(0); setSelected([]); }} onBack={()=>setStage(2)} blendHex={blendHex} audioToggle={<AudioToggle />} />}
       </div>
     </div>
   );
